Image/sprite loader and animation viewer
Browserquest sprite descriptions are stored in JSON files. This
code provides a rudimentary way of loading sprite and image data
and previewing animations using Swing.

We're using [Jerkson] wrapped it in a thin `JsonFormat` trait to
parse JSON.

The `AnimationViewer` class provides the main entry point. Run with:

    sbt 'run <spritename>'

for example:

    sbt 'run leatherarmor'

There's also a bunch of math code like `Vec` and `AxisBox` that
we're not using very much yet.