Minecraft-style generation and caching of chunks of the game world
Added `scalaquest.world` containing a basic terrain generation
mechanism along similar lines to that used in Minecraft.

The `World` is divided into 16x16 tile `Chunk`s. Chunks are
generated on-demand as they come on-screen, and are then cached
for later use. This gives us a really large streaming world that
can be modified by the player.

The `WorldTest` class provides a really simple scrolling view of the
world. The `World` logs the generation of new chunks as they scroll
on-screen.